:recycle: refactor: ProductsCart component and add useProductsCart hook
